LayoutTests/http/tests/security/contentSecurityPolicy/1.1/securitypolicy-isactive-expected.txt

PASS document.SecurityPolicy.active is false when no policy exists.
PASS document.SecurityPolicy.active is true when policy exists.
